+++ repository/sqlite/users.go
@@ -1,4 +1,4 @@
-package repository
+package sqlite
 
 import (
 	"database/sql"
